refactor: InputCheck 에서 InputValidator로 클래스명을 변경한다.
